결제API, 관리자(결제관리), 마이페이지(결제내역),
상엽

+++ IcuProject/src/main/java/com/kh/icu/party/model/vo/Party.java
@@ -21,7 +21,7 @@ public class Party {
 	private Date startDate;
 	private Date endDate;
 	private int dayPrice;
-	private int account;
+	private Long account;
 	private int status;
 	private int leftDate;
 	private int joinNum;
